requirements.txt
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
pytest==8.3.3
pytest-cov==6.0.0